fix after first review

+++ database/migrations/2021_04_28_131028_create_wallets_table.php
@@ -16,7 +16,7 @@ class CreateWalletsTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
-            $table->double('current_balance');
+            $table->bigInteger('current_balance');
             $table->string('hash');
         });
     }
